requirements.txt
openai-whisper
pydub
vosk
transformers
wave